룸 관련 오류 수정 테스트 필요
