+++ app/src/main/java/com/android/hanple/ui/search/InitLoadFragment.kt
@@ -1,23 +1,38 @@
 package com.android.hanple.ui.search
 
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.whenStarted
 import com.android.hanple.R
+import com.android.hanple.Room.RecommendDataBase
+import com.android.hanple.Room.RecommendPlace
+import com.android.hanple.Room.recommendPlaceGoogleID
 import com.android.hanple.databinding.FragmentInitLoadBinding
 import com.android.hanple.databinding.FragmentSearchBinding
+import com.android.hanple.viewmodel.SearchViewModel
+import com.android.hanple.viewmodel.SearchViewModelFactory
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
+import kotlin.random.Random
 
 
 class InitLoadFragment : Fragment() {
     private var _binding : FragmentInitLoadBinding? = null
     private val binding get() = _binding!!
     private var time: Int = 0
+    private val viewModel by lazy {
+        ViewModelProvider(requireActivity(), SearchViewModelFactory())[SearchViewModel::class.java]
+    }
+    private val recommendDAO by lazy {
+        RecommendDataBase.getMyRecommendPlaceDataBase(requireContext()).getMyRecommendPlaceDAO()
+    }
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -25,7 +40,6 @@ class InitLoadFragment : Fragment() {
         _binding = FragmentInitLoadBinding.inflate(layoutInflater)
         return binding.root
     }
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         initView()
@@ -39,10 +53,11 @@ class InitLoadFragment : Fragment() {
     private fun initView(){
         lifecycleScope.launch {
             whenStarted{
+                setRecommendPlace()
                 while (true){
                     delay(1000)
                     time += 1
-                    if(time == 2){
+                    if(time == 3){
                         time = 0
                         val searchFragment = SearchFragment()
                         val transaction = parentFragmentManager.beginTransaction()
@@ -53,5 +68,23 @@ class InitLoadFragment : Fragment() {
                 }
             }
         }
+    }
+    private fun randomNumberPlace(): List<Int> {
+        val edge = recommendPlaceGoogleID.size
+        val list = mutableListOf<Int>()
+        var number: Int = 0
+        while (list.size < 5) {
+            number = Random.nextInt(edge) + 1
+            if (list.contains(number))
+                continue
+            else
+                list.add(number)
+        }
+        return list
+    }
+
+    private fun setRecommendPlace() {
+        val list = randomNumberPlace()
+        viewModel.getRecommendPlace(list, recommendDAO)
     }
 }
